third_party/WebKit/LayoutTests/svg/animations/svgenum-animation-6-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Test ColorMatrixType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_SATURATE
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_SATURATE
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_HUEROTATE
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_HUEROTATE
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_LUMINANCETOALPHA
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_LUMINANCETOALPHA
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.animVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS colorMatrix.type.baseVal is SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX
PASS successfullyParsed is true

TEST COMPLETE
